Changes made for Bug # 17204.
The Browser header was not proper when bulk operation was integrated in ClinPortal. This was due to non generic values being put in bulkoperations-tiles.defs.xml file.
The values for properties in bulkoperations-tiles.defs.xml should be obtained from the host application and should be replaced during deploy time.
To perform this task of replacing values an ant task is written.

Code Reviewed By : Shital L

Code Reviewed On : May 24 2010

Testcase : NA

SVN-Revision: 275